Commiting from makefile for new branch

+++ util.h
@@ -10,5 +10,5 @@
 void util_init();
 void util_ledonoff(unsigned char);
 uint8_t util_read_calib_byte(uint8_t);
-uint8_t util_wait_ms(int);
+int util_wait_ms(int);
 char * util_ifmt(int, uint8_t);
